refactor: separate unit tests from integration tests in pkg/triggers
- Created filename_integration_test.go for registry-based tests
- Created registry_integration_test.go for catchall registry tests
- Removed registry tests from unit test files
- All trigger unit tests now use mocks only
- Integration tests handle real registry interactions
